+++ weaver/processes/builtin/jsonarray2netcdf.py
@@ -7,10 +7,11 @@ import json
 import logging
 import os
 import sys
-from typing import Any, AnyStr
+from typing import AnyStr
 
 import six
-from six.moves.urllib.parse import urlparse
+
+from weaver.processes.builtin.utils import _is_netcdf_url
 
 if six.PY3:
     from tempfile import TemporaryDirectory
@@ -24,7 +25,6 @@ sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.dirname(CUR_DIR))))
 
 # place weaver specific imports after sys path fixing to ensure they are found from external call
 # pylint: disable=C0413,wrong-import-order
-from weaver.formats import get_extension, CONTENT_TYPE_APP_NETCDF  # isort:skip # noqa: E402
 from weaver.utils import fetch_file  # isort:skip # noqa: E402
 
 PACKAGE_NAME = os.path.split(os.path.splitext(__file__)[0])[-1]
@@ -38,15 +38,6 @@ LOGGER.setLevel(logging.INFO)
 __version__ = "1.0"
 __title__ = "JSON array to NetCDF"
 __abstract__ = __doc__  # NOTE: '__doc__' is fetched directly, this is mostly to be informative
-
-
-def _is_netcdf_url(url):
-    # type: (Any) -> bool
-    if not isinstance(url, six.string_types):
-        return False
-    if urlparse(url).scheme == "":
-        return False
-    return os.path.splitext(url)[-1] == get_extension(CONTENT_TYPE_APP_NETCDF)
 
 
 def j2n(json_reference, output_dir):
