+++ weaver/processes/builtin/metalink2netcdf.py
@@ -10,7 +10,6 @@ from typing import AnyStr
 
 import six
 from lxml import etree
-from six.moves.urllib.parse import urlparse
 
 if six.PY3:
     from tempfile import TemporaryDirectory
